feat: givenNameType support annimal
