+++ dict/words.json
@@ -1,10 +1,11 @@
 {
   "normal": "晨航乐星逸宁然瑾岚轩希昕瑜澜安佑远翊清诚皓义文涵明柯卓昱奇宏阳秋佳彦景霄舒彬晗乔依雅晓元南修子桦荣睿善宸峥利泽运辰腾竹铭琪林昀志翰懿亦凡殷哲书嘉仪旭梓毓全骏宇潇祺康信云永君少凌萧伊悠衡霖扬",  
-  "female": "芳娜美艳玲琳丽霞颖静婷茹露诗韵思俪莉柔梦沐霏依芊婉晴珍语珂绮汶怡樱雯佩慧雪娇曼灵珊咏彤芷芸菁蔚英蕊洁可岑雨霜彩穗绣嫣姬萱茜琦瑶芮茵璇惠心欣馨歆燕尧倩莹俐妍梅萍媛妮妙敏",
+  "female": "芳娜美艳玲琳丽霞颖静婷茹露诗韵思俪莉柔梦沐霏依芊婉晴珍语珂绮汶怡樱雯佩慧雪娇曼灵珊咏彤芷芸菁蔚英蕊洁可岑雨霜彩穗绣嫣姬萱茜琦瑶芮茵璇惠心欣馨歆燕凤尧倩莹俐妍梅萍媛妮妙敏",
   "male": "伟强浩俊凯龙锋志宇博旭豪昊涛军杰超华勇祥峰诚炜熙锐翔峻朗弘天震瑞斌劲良泽源勋纶智炎德昆佑宾克庆维鼎野翼奕伦承鑫昭贺光展鸿冠威亮毅辉卫健韬东富",
   "wood": "松柏楠柳桦森桂槿榆桐棠枫栋栩杉梧",
   "fire": "炎炜烽烁焕煜灿炫烨曦晴烈焱晖晗旭",
   "earth": "坤岳峰岩岱峥磊垚境岚城培均峻基",
   "metal": "锋锐钧铄铭钟钰鑫铠锴钦锡镇铮锦",
-  "water": "涛淼淇波海泉渊涵洁源泷泽滨泓沐"
+  "water": "涛淼淇波海泉渊涵洁源泷泽滨泓沐",
+  "animal": "龙麟凤鹿鹤鱼燕鹰蝶蝉雀鲸彪象兔"
 }
