+++ docs/en/givenNameType.md
@@ -9,3 +9,4 @@
 |        | water     | 水   | 涛、淼、润、波、海... |
 |        | fire      | 火    | 炎、炜、烽、烁、焕... |
 |        | earth     | 土   | 坤、岳、峰、岩、岱... |
+| Others  | animal  | Animal  | 龙、麟、凤、鹿、鹤... |
