+++ src/utils/default.ts
@@ -17,3 +17,9 @@ export const DEFAULT_SURNAME_ALGORITHM: Algorithm = 'weight';
  * 名字属性的优先级顺序，决定如何筛选名字
  */
 export const GIVEN_NAME_TYPE_PRIORITY = ['elements', 'gender'] as const;
+
+/**
+ * @private
+ * 以下类型的名字最好只包含一个字
+ */
+export const GIVEN_NAME_INCLUDE_ONE_CHARACTER_TYPE = ['animal'];
